requirements.txt
bcrypt==3.1.7
Flask==1.1.1
mysql-connector-python==8.0.19
pytest==5.3.5
